Pass -Cstrip=debuginfo when emitting relocations
A change included in 1.80.1 (or a dependent crate) is causing
`-Wl,--strip-all` to be passed to linker commands when we are trying to
emit relocations.  The linker emits an error when it is told to remove
all symbols while also emitting more symbols.  This change will tell
Rust to only strip debug symbols from binaries, which is compatible with
emitting relocations.

Test: ./tools/build.py --llvm-linkage shared --lto thin --emit-relocs
Bug: 361370910
Change-Id: I2b57ad8eda9d0f798ee4ddab38ffeb65086652c3