project/qemu: Avoid potential exception on shutdown
Since reboot commands call shutdown and shutdown is also called at
the end of a test run (regardless of the mode of termination), it
is possible for shutdown to be called twice. The cleanup step will
fail the second time around if the temporary files have already
been removed.

This is a problem because shutdown can be on the stack frame for
arbitrarily long time if debug on error is set and an error occurs.

Bug: None
Test: manually, run qemu-generic-arm64-test-debug via the run script
      and cause it to terminate while shutdown is being called from a
      reboot command.
Change-Id: I650e9b0db5fa88b5c66e98a0c8cafa2323c321d4