Enable access of raw pointer for shared variables
The current mechanism of SharedVariables in Ditto works by passing
variables by value. This works fine for most of the cases, but is a
limitation for those types that cannot be copied.
For those cases, the simplest solution is to let the SharedVariables
utility class to expose their raw pointer.

Test: cmake tests
Bug: 365781383
Change-Id: I8251d6a1d8f7c77bd6fb42a7e9644a08ab233ef5
Signed-off-by: Alessio Balsini <balsini@google.com>