Source tools/buildbot-utils.sh from tools/run-gtests.sh.
tools/buildbot-utils.sh sets up some environment variables that are
necessary in order to run gtests on QEMU (which current approach used
for testing riscv64 on LUCI).

Without this import gtests failed with the following error:
    find: ‘/apex/com.android.art/bin/art’: No such file or directory
(although the error was not detected and the test run was green on the
LUCI console).

Test: launched custom LUCI job with this patch:
    led get-builder art/ci/qemu-riscv64-ndebug \
        | led edit-gerrit-cl https://r.android.com/3282595 \
        | led launch
    (observed that gtests run, some of them fail)

Change-Id: I3bde4b5f21defa5ce118c9e2df780ee0970f4680