[gbl] move decompression to libgbl
In order to move Android boot logic into libgbl, we'll need the
decompression routines to be present there as well. This does add a more
explicit dependency on allocation, but it was already there since libavb
needs alloc, so this shouldn't make it any harder to integrate libgbl
than it already was.

This CL also cleans up some outdated comments on alloc, and makes
decompression non-arch-dependent. I had originally recommended making
this code arch-dependent but it turns out this isn't great for unit
tests (since the tests only run on the host arch), and since there
isn't actually anything specific to ARM in this code it seemed better to
go back to arch-independence.

Bug: b/363074091
Change-Id: Ie9caa2750ffd051b9ca083432415f73a36ac5be2