Autoupdate of 2 submodules
*frameworks/base updated to 9962460830cf6e48b6d2bb0842edb0c4a6741a61
Merge "Restrict use of framework-annotations" into main
*frameworks/libs/modules-utils updated to 6f3170d0543c8214dd89382b145b4e30d424372c
Restrict use of framework-annotations

This filegroup target should only be used when the raw srcs are needed.
Module targets can use `framework-annotations-lib` as a libs reference,
if necessary. Otherwise, the annotations are bundled directly with the
framework.

This refactor is in anticipation of bundling keep rules with the
framework-annotations-lib target, instead of defining global keep rules.

Bug: 275175207
Test: m
Flag: EXEMPT refactor
Change-Id: I266501244ca372de4357999d485e8d1cfe0a2d9b